backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.30.0
chromadb>=0.5.3
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0
sqlalchemy>=2.0.0
pyodbc>=5.1.0
alembic>=1.13.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>

# Ai FastAPI + Uvicorn pentru server HTTP.

# OpenAI pentru LLM.

# ChromaDB pentru vector store.

# python-dotenv pentru .env.

# pydantic v2 pentru modele.

# tqdm pentru progres la ingest.

# loguru pentru logare frumoasă.

# SQLAlchemy + pyodbc pentru SQL Server.

# Alembic pentru migrații DB.

# httpx + pytest pentru testare.

# python-jose + passlib pentru auth (opțional).

# python-dateutil pentru manipulare date.